final fix

+++ app/models/currency.rb
@@ -1,16 +1,18 @@
 class Currency < ApplicationRecord
   class << self
     def update_currencies
-      destroy_all
-
       require 'net/http'
 
       url = URI.parse('http://bank.gov.ua/NBUStatService/v1/statdirectory/exchange?json')
       req = Net::HTTP::Get.new(url.to_s)
       res = Net::HTTP.start(url.host, url.port) { |http| http.request(req) }
+      body = JSON.parse(res.body)
 
-      JSON.parse(res.body).each do |item|
-        create(name: item[:cc], rate: item[:rate])
+      destroy_all if body.present?
+
+      create(name: 'Українська гривня', rate: '1')
+      body.each do |item|
+        create(name: item['txt'], rate: item['rate'])
       end
     end
   end
